Make requests asynchronously
This commit changes requests from sync to async. This is especially
useful when using Emacs to both make the HTTP request (with `walkman`)
and with a terminal for local development. In such a scenario, we can't
let Emacs hang on the HTTP request, otherwise we can't access the
terminal.

To make this change I needed to change `call-process` to `make-process`.
Since the request now happens asynchronously, I needed to change the way
the response is parsed, which now happens inside a sentinel (which is
called once the process created by `make-process` ends).

Closes #7.