Change regex in form parsing to only replace first occurrence (#21)
Currently when parsing `:FORM` headers, we're replacing every occurrence
of `: ` with `=`. However, when the form header value also contains `:`,
we're wrongly replacing it with `=` as well – for instance, when
parsing an header that contains JSON object, we're replacing all of the
`:` in the key-value syntax with `=`, thus producing invalid JSON.

This commit fixes that by only changing the first occurrence. This is
done following the advice on the documentation for
`replace-regexp-in-string` function.